- fixed linter by moving retry.go to its own package
