+++ config/config.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/DIMO-Network/edge-network/internal/util"
+	"github.com/DIMO-Network/edge-network/internal/util/retry"
 	"github.com/rs/zerolog"
 
 	"github.com/DIMO-Network/shared"
@@ -109,7 +109,7 @@ func ReadConfig(logger zerolog.Logger, configFiles embed.FS, configURL, configFi
 	// Get secrets from remote config
 	remoteConfigPathOnDisk := "/opt/autopi/remote-config.json"
 	// Retry for about 1 hour
-	remoteConfig, err := util.Retry[Config](11, 1*time.Second, logger, func() (interface{}, error) {
+	remoteConfig, err := retry.Retry[Config](11, 1*time.Second, logger, func() (interface{}, error) {
 		return GetRemoteConfig(configURL, remoteConfigPathOnDisk)
 	})
