+++ internal/util/retry/retry.go
@@ -1,4 +1,4 @@
-package util
+package retry
 
 import (
 	"time"
